Add download-file shared file
