+++ application/ui/src/platform/download-file.ts
@@ -1,17 +1,8 @@
 // Copyright (C) 2026 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
+import { downloadViaAnchor } from './download-file.shared';
+
 export const downloadFile = (url: string, name?: string): void => {
-    const link = document.createElement('a');
-
-    link.href = url;
-    if (name !== undefined) {
-        link.download = name;
-    }
-    link.hidden = true;
-    link.click();
-
-    if (url.startsWith('blob:')) {
-        URL.revokeObjectURL(url);
-    }
+    downloadViaAnchor(url, name);
 };
